LayoutTests/fast/filesystem/workers/sync-operations.html

<html>
<head>
<script src="../resources/fs-worker-common.js"></script>
<script src="../resources/fs-worker-test-pre.js"></script>
</head>
<body>
<p id="description">This tests running multiple sync filesystem operations in Worker.</p>
<div id="console"></div>
<script>
startWorker('../resources/sync-operations.js');
</script>
<script src="../resources/fs-worker-test-post.js"></script>
</body>
</html>
